+++ __tests__/components/header.test.js
@@ -1,15 +1,22 @@
 import { useSession } from "next-auth/react"
 import { render } from "@testing-library/react"
 
+import { session } from "../../src/mocks/data"
 import Header from "../../src/components/header"
 
-// https://github.com/nextauthjs/next-auth/issues/775
-jest.mock("next-auth/react", () => ({
-  useSession: jest.fn(),
-}))
+describe("when header renders", () => {
+  it("does not show authenticated user if session is not valid", () => {
+    useSession.mockImplementation(() => [false, false])
+    const { container } = render(<Header />)
+    expect(container).toMatchSnapshot()
+  })
 
-it("renders header", () => {
-  useSession.mockReturnValueOnce([false, false])
-  const { container } = render(<Header />)
-  expect(container).toMatchSnapshot()
+  it("does show authenticated user if session is valid", () => {
+    useSession.mockImplementation(() => ({
+      data: session,
+      status: "authenticated",
+    }))
+    const { container } = render(<Header />)
+    expect(container).toMatchSnapshot()
+  })
 })